HLNA-3 fix review comments

+++ src/main/resources/schema.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS crew
 (
-    id   BIGINT AUTO_INCREMENT,
+    id   INT AUTO_INCREMENT,
     name VARCHAR(50) NOT NULL,
 
     CONSTRAINT crew_PK PRIMARY KEY (id)
@@ -8,7 +8,7 @@ CREATE TABLE IF NOT EXISTS crew
 
 CREATE TABLE IF NOT EXISTS airplane
 (
-    id               BIGINT AUTO_INCREMENT,
+    id               INT AUTO_INCREMENT,
     code_name        VARCHAR(50) NOT NULL,
     model            VARCHAR(50) NOT NULL,
     manufacture_date DATE        NOT NULL,
@@ -22,13 +22,13 @@ CREATE TABLE IF NOT EXISTS airplane
 
 CREATE TABLE IF NOT EXISTS crew_member
 (
-    id          BIGINT AUTO_INCREMENT,
+    id          INT AUTO_INCREMENT,
     first_name  VARCHAR(50) NOT NULL,
     last_name   VARCHAR(50) NOT NULL,
     position    VARCHAR(50) NOT NULL,
     birthday    DATE        NOT NULL,
     citizenship VARCHAR(25),
-    crew_id     BIGINT      NOT NULL,
+    crew_id     INT      NOT NULL,
 
     CONSTRAINT crew_member_PK PRIMARY KEY (id),
     CONSTRAINT crew_member_crew_FK FOREIGN KEY (crew_id) REFERENCES crew (id)
